Improve test

+++ test/load_command_version_3_default.test.ts
@@ -1,30 +1,33 @@
-import { createClient } from '@/groonga-stdio-client'
+import { createClient, GroongaStdioClient } from '@/groonga-stdio-client'
 import { mkdir, rimraf, createOptions, shutdown } from './test_utils'
 import path from 'path'
 
-jest.setTimeout(30000)
-
 describe('GroongaStdioClient', () => {
-  const db_dir = path.join(__dirname, 'db_load')
+  let db_dir: string | undefined
+  let client: GroongaStdioClient | undefined
 
-  beforeAll(() => {
-    rimraf(db_dir)
-    mkdir(db_dir)
+  beforeEach(() => {
+    db_dir = undefined
+    client = undefined
   })
 
-  afterAll(() => {
-    return new Promise((resolve) => {
-      setTimeout(() => {
-        rimraf(db_dir)
-        resolve()
-      }, 800)
-    })
+  afterEach(async () => {
+    if (client) {
+      await shutdown(client).catch(() => null)
+    }
+    if (db_dir) {
+      rimraf(db_dir)
+    }
   })
 
   test('load/command_version/3/default', async () => {
-    const db_path = path.join(db_dir, `load.db`)
+    db_dir = path.join(__dirname, 'temp.load_command_version_3')
+    const db_path = path.join(db_dir, `db`)
+    rimraf(db_dir)
+    mkdir(db_dir)
+
     const opts = createOptions()
-    const client = createClient(db_path, opts)
+    client = createClient(db_path, opts)
 
     expect(client).not.toBeUndefined()
 
@@ -32,21 +35,17 @@ describe('GroongaStdioClient', () => {
       return
     }
 
-    try {
-      const r1 = await client.commandAsync('table_create Memos TABLE_NO_KEY')
-      expect(r1).toBe(true)
+    const r1 = await client.commandAsync('table_create Memos TABLE_NO_KEY')
+    expect(r1).toBe(true)
 
-      const r2 = await client.commandAsync('column_create Memos value COLUMN_SCALAR Int8')
-      expect(r2).toBe(true)
+    const r2 = await client.commandAsync('column_create Memos value COLUMN_SCALAR Int8')
+    expect(r2).toBe(true)
 
-      const r3 = await client.commandAsync('load --table Memos --command_version 3', {
-        values: JSON.stringify([{ value: 1 }, { value: 2 }]),
-      })
-      expect(r3).toEqual({
-        n_loaded_records: 2,
-      })
-    } finally {
-      await shutdown(client)
-    }
+    const r3 = await client.commandAsync('load --table Memos --command_version 3', {
+      values: JSON.stringify([{ value: 1 }, { value: 2 }]),
+    })
+    expect(r3).toEqual({
+      n_loaded_records: 2,
+    })
   })
 })
